+++ packages/database-devtools/src/inspectors/sqlite/index.ts
@@ -28,5 +28,3 @@ export {
   sqliteInspectorDefinition,
 };
 export { configureSqliteWasm } from './sqlModule';
-
-registerSqliteInspector();